Don't use long, but size_t for bytecodeIndex

+++ src/vmobjects/VMFrame.h
@@ -79,7 +79,7 @@ private:
     GCFrame*  previousFrame;
     GCFrame*  context;
     GCMethod* method;
-    long bytecodeIndex;
+    size_t    bytecodeIndex;
     gc_oop_t* arguments;
     gc_oop_t* locals;
     gc_oop_t* stack_ptr;
